- Modified the makefiles to work recursively - Added makefiles usage instructions to README.txt - fixed a few files to make them compile on Cygwin - Added Build folders to ignore list

+++ runtime/zorba.cpp
@@ -16,8 +16,6 @@
 #include "../store/api/item_factory.h"
 
 
-
-
 using namespace std;
 namespace xqp {
 
@@ -83,7 +81,7 @@ zorba::getZorbaForCurrentThread()
 	std::map<uint64_t, zorba*>::iterator	it_zorba;
 
 	pthread_mutex_lock(&global_zorbas_mutex);
-	it_zorba = global_zorbas.find((uint64_t)pthread_self());
+	it_zorba = global_zorbas.find((uint64_t)(uintptr_t)pthread_self());
 	if(it_zorba == global_zorbas.end())
 	{///not found, big error
 		pthread_mutex_unlock(&global_zorbas_mutex);
@@ -104,7 +102,7 @@ zorba::allocateZorbaForNewThread()
 	new_zorba = new zorba();
 
 	pthread_mutex_lock(&global_zorbas_mutex);
-	global_zorbas[(uint64_t)pthread_self()] = new_zorba;
+	global_zorbas[(uint64_t)(uintptr_t)pthread_self()] = new_zorba;
 	pthread_mutex_unlock(&global_zorbas_mutex);
 
 	return new_zorba;
@@ -115,7 +113,7 @@ void
 zorba::destroyZorbaForCurrentThread()//when ending the thread
 {
 	pthread_mutex_lock(&global_zorbas_mutex);
-	global_zorbas.erase((uint64_t)pthread_self());
+	global_zorbas.erase((uint64_t)(uintptr_t)pthread_self());
 	pthread_mutex_unlock(&global_zorbas_mutex);
 }
